+++ frontend/controllers/SecuredController.php
@@ -7,7 +7,7 @@ use Yii;
 use yii\filters\AccessControl;
 use yii\web\Controller;
 
-class BaseController extends Controller
+abstract class SecuredController extends Controller
 {
     public function behaviors()
     {
@@ -23,19 +23,4 @@ class BaseController extends Controller
             ]
         ];
     }
-
-    public function getProfile()
-    {
-        if ($id = \Yii::$app->user->getId()) {
-            $user = User::findOne($id);
-            return $user->name;
-        }
-    }
-
-    public function actionLogout()
-    {
-        Yii::$app->user->logout();
-        return $this->goHome();
-    }
-
 }
