+++ frontend/controllers/SiteController.php
@@ -18,7 +18,7 @@ use frontend\models\ContactForm;
 /**
  * Site controller
  */
-class SiteController extends BaseController
+class SiteController extends SecuredController
 {
     /**
      * {@inheritdoc}
